Remove cyclical dependency in router<->webpush
