+++ client/js/router.js
@@ -142,4 +142,18 @@ function switchToChannel(channel) {
 	return navigate("RoutedChat", {id: channel.id});
 }
 
+if ("serviceWorker" in navigator) {
+	navigator.serviceWorker.addEventListener("message", (event) => {
+		if (event.data && event.data.type === "open") {
+			const id = parseInt(event.data.channel.substr(5), 10); // remove "chan-" prefix
+
+			const channelTarget = store.getters.findChannel(id);
+
+			if (channelTarget) {
+				switchToChannel(channelTarget.channel);
+			}
+		}
+	});
+}
+
 export {initialize, router, navigate, switchToChannel};
